+++ assets/components/common/BaseTimer.tsx
@@ -1,4 +1,6 @@
 import React, { useState, useEffect, useRef, useImperativeHandle, forwardRef } from 'react';
+import Sound from 'react-native-sound';
+import MetronomeModule from "react-native-metronome-module";
 
 /** @Assets */
 import { CommonStyles } from 'styles';
@@ -9,16 +11,32 @@ import { BaseText } from './BaseText';
 /** */
 
 import { getClock } from '../../../src/utils';
+import { useSelector } from 'react-redux';
+import { RootState } from '../../../src/app/store';
+import { getSettingsState } from '../../../src/app/slices/settingsSlice';
+
+const CYCLE = 25;
+const LAST_SECONDS = 10;
 
 interface BaseTimerProps{
 }
 
+Sound.setCategory('Playback');
+
 function BaseTimer(props: BaseTimerProps, ref: any){
 
+  const settings = useSelector((rootState: RootState) => getSettingsState(rootState));
   const [basis, setBasis] = useState(0);
   const [timer, setTimer] = useState(0);
   const [timerDisp, setTimerDisp] = useState('00:00');
   const intervalId = useRef(0);
+
+  const longding = useRef(new Sound('longnotif.mp3', Sound.MAIN_BUNDLE, (error) => {
+    if (error) {
+      console.log('failed to load the sound', error);
+      return;
+    }
+  }));
 
   useEffect(() => {
     if (basis)
@@ -35,17 +53,24 @@ function BaseTimer(props: BaseTimerProps, ref: any){
     if (basis && timer) {
       const toDisp = Math.floor((timer - basis) / 1000)
       setTimerDisp(getClock(toDisp));
+
+      if(toDisp >= CYCLE && toDisp % CYCLE === 0){        
+        longding.current.play();
+      }
     }
   }, [timer]);
 
   function handleStart(){
     const t = new Date();
     setBasis(t.valueOf());
+    MetronomeModule.setBPM(settings.bpm);
+    MetronomeModule.start();
   }
 
   function handleStop(){
     setTimerDisp('00:00');
     clearInterval(intervalId.current);
+    MetronomeModule.stop();
   }
 
   useImperativeHandle(ref, () => ({
