+++ package.json
@@ -35,6 +35,7 @@
     "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "^2.12.1",
     "react-native-html-to-pdf": "^0.12.0",
+    "react-native-metronome-module": "^2.3.0",
     "react-native-reanimated": "^3.3.0",
     "react-native-root-toast": "^3.5.1",
     "react-native-safe-area-context": "^4.7.2",
